Reorganized rice module
