+++ src/java/growthcraft/rice/ClientProxy.java
@@ -1,5 +1,8 @@
 package growthcraft.rice;
 
+import growthcraft.rice.renderer.RenderRice;
+import growthcraft.rice.renderer.RenderPaddy;
+
 import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class ClientProxy extends CommonProxy
