+++ src/java/growthcraft/rice/block/IPaddyCrop.java
@@ -1,3 +1,3 @@
-package growthcraft.rice;
+package growthcraft.rice.block;
 
 public interface IPaddyCrop {}
