+++ src/java/growthcraft/rice/item/ItemRice.java
@@ -1,14 +1,16 @@
-package growthcraft.rice;
+package growthcraft.rice.item;
 
 import growthcraft.core.GrowthCraftCore;
+import growthcraft.rice.GrowthCraftRice;
+
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 public class ItemRice extends Item
 {
