+++ src/java/growthcraft/rice/renderer/RenderPaddy.java
@@ -1,6 +1,12 @@
-package growthcraft.rice;
+package growthcraft.rice.renderer;
 
 import growthcraft.core.Utils;
+import growthcraft.rice.ClientProxy;
+import growthcraft.rice.GrowthCraftRice;
+import growthcraft.rice.block.BlockPaddy;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.EntityRenderer;
 import net.minecraft.client.renderer.RenderBlocks;
@@ -8,11 +14,7 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-
 import org.lwjgl.opengl.GL11;
-
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderPaddy implements ISimpleBlockRenderingHandler
 {
