+++ src/java/growthcraft/rice/renderer/RenderRice.java
@@ -1,13 +1,15 @@
-package growthcraft.rice;
+package growthcraft.rice.renderer;
 
+import growthcraft.rice.block.BlockRice;
+
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
+import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.renderer.EntityRenderer;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.IIcon;
 import net.minecraft.world.IBlockAccess;
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
-import cpw.mods.fml.client.registry.RenderingRegistry;
 
 public class RenderRice implements ISimpleBlockRenderingHandler
 {
